added check if the directory exists

+++ online/junk/Remove_Junk_Directories.ps1
@@ -6,7 +6,10 @@ $PSDir = Split-Path -Parent $MyInvocation.MyCommand.Path
 $JunkPath = (Join-Path $PSDir "\directories.lst")
 $Items = Get-Content -Path $tempPath
 foreach($Item in $Items){
-    Write-Host '[Directory] Removing Junk at:' $Item -ForegroundColor Cyan
-    Remove-Item -Path $Item -Recurse -Force -ErrorAction SilentlyContinue -Confirm:$false}
+    if (Test-Path $Item){
+        Write-Host '[Directory] Removing Junk at:' $Item -ForegroundColor Cyan
+        Remove-Item -Path $Item -Recurse -Force -ErrorAction SilentlyContinue -Confirm:$false}
+    }
+    
 # Finally delete this script
 Remove-Item -Path $MyInvocation.MyCommand.Source -Force -ErrorAction SilentlyContinue
